Moved PostProcessingOptions to its own script
