+++ Source/Utility/PostProcessingOptions.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 967d5e4e2e041f74684c53b2b44196e7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
